feat(common-circular-dependency): remove circular dependencies on common package
affects: @zetapush/common
